refactor: add comments and tags to trait code
