+++ Assets/Scripts/Data/GoopTraits/GoopTraitDataScalar.cs
@@ -3,16 +3,40 @@ using UnityEngine;
 
 namespace GoopGame.Data
 {
+    /// <summary>
+    /// Float implementation of <seealso cref="GoopTraitData{T}"/>
+    /// </summary>
     [CreateAssetMenu(
         fileName = "GoopScalarTrait", 
         menuName = "GoopGame/Goops/Traits/Create new Goop Scalar Trait"
         )]
     public class GoopTraitDataScalar : GoopTraitData<float>
     {
+        /// <summary>
+        /// Lowest default value.
+        /// </summary>
         public float DefaultLowerBound;
+
+        /// <summary>
+        /// Highest default value.
+        /// </summary>
         public float DefaultUpperBound;
+
+        /// <summary>
+        /// The probability of generating an initial value within the 
+        /// default bounds.
+        /// </summary>
         public ProbabilityCurve DefaultBoundsProbability;
+
+        /// <summary>
+        /// The maximum mutation range.
+        /// </summary>
         public float MutationRange;
+
+        /// <summary>
+        /// The probability of generating a value from -MutationRange to 
+        /// MutationRange.
+        /// </summary>
         public ProbabilityCurve MutationRangeProbability;
 
         public override float GenerateRandomValue()
@@ -31,6 +55,7 @@ namespace GoopGame.Data
             GoopWeightStruct weights2
             )
         {
+            //Calculate parent lerp value, and combine.
             float parentCombinationRatio = 
                 ParentsLerpProbability.GetValue(Random.Range(0f, 1f));
             float combinedValue = Mathf.Lerp(
@@ -40,12 +65,14 @@ namespace GoopGame.Data
                 weights1, weights2, parentCombinationRatio
                 );
 
+            //Calculate biases.
             foreach (var bias in EvolutionWeights)
             {
                 if (bias.IsBiasApplied(combinedStruct))
                     ApplyBias(ref combinedValue, bias);
             }
 
+            //Single mutation step.
             DoMutation(ref combinedValue);
 
             return combinedValue;
@@ -55,9 +82,11 @@ namespace GoopGame.Data
             float value, GoopWeightStruct weights
             )
         {
+            //Store values on stack, to be passed by reference.
             float value1 = value;
             float value2 = value;
 
+            //Evaluate biases.
             foreach(var bias in EvolutionWeights)
             {
                 if (bias.IsBiasApplied(weights))
@@ -66,6 +95,7 @@ namespace GoopGame.Data
                     ApplyBias(ref value2, bias);
             }
 
+            //Single mutation step.
             DoMutation(ref value1);
             DoMutation(ref value2);
 
@@ -74,10 +104,13 @@ namespace GoopGame.Data
 
         public override void ApplyBias(ref float value, EvolutionBias bias)
         {
+            //Get delta.
             float delta = bias.BiasTarget - value;
+            //Clamp within maximum intensity.
             delta = Mathf.Clamp(
                 delta, -bias.MaximumBiasIntensity, bias.MaximumBiasIntensity
                 );
+            //Multiply delta by weighted random value.
             delta = Mathf.Lerp(
                 0f, delta, 
                 bias.BiasIntensityProbability.Evaluate(Random.Range(0f, 1f))
@@ -87,10 +120,12 @@ namespace GoopGame.Data
 
         public override void DoMutation(ref float value)
         {
+            //Calculate intensity
             float mutationIntensity = 
                 MutationRangeProbability.GetValue(Random.Range(0f, 1f));
+            //Calculate mutation
             float mutation = 
-                Mathf.Lerp(0f, MutationRange, mutationIntensity);
+                Mathf.Lerp(-MutationRange, MutationRange, mutationIntensity);
             value += mutation;
         }
     }
