+++ Assets/Scripts/Data/GoopTraits/TypeWeightPair.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: da99145cc37b00f47b79a5b695b1c001
