+++ Assets/Scripts/Engine/Goop.cs
@@ -5,11 +5,22 @@ namespace GoopGame.Engine
 {
     public class Goop : MonoBehaviour
     {
+        /// <summary>
+        /// Static stat data.
+        /// </summary>
         [SerializeField]
         private GoopStatData _hungerData, _temperatureData, 
             _moodData, _energyData;
+
+        /// <summary>
+        /// Static float trait data.
+        /// </summary>
         [SerializeField]
         private GoopTraitDataScalar _sizeData, _speedData;
+
+        /// <summary>
+        /// Static color trait data.
+        /// </summary>
         [SerializeField]
         private GoopTraitDataColor _colorData;
 
@@ -18,6 +29,9 @@ namespace GoopGame.Engine
         /// </summary>
         public GoopStats Stats { get; private set; }
 
+        /// <summary>
+        /// The immutable stats of a goop instance.
+        /// </summary>
         public GoopTraits Traits { get; private set; }
 
         /// <summary>
@@ -26,8 +40,14 @@ namespace GoopGame.Engine
         [field: SerializeField]
         public BaseState State { get; private set; }
 
+        /// <summary>
+        /// Determines if the goop initializes itself on Start()
+        /// </summary>
         private bool _initialized = false;
 
+        /// <summary>
+        /// Sets FSM state of goop.
+        /// </summary>
         public void SetState(BaseState state)
         {
             State.ExitState(this); // Exit the current state
@@ -35,11 +55,19 @@ namespace GoopGame.Engine
             State.EnterState(this); // Enter the new state
         }
 
+        /// <summary>
+        /// Sets the traits of this goop. Should be called before
+        /// initializing goop to avoid generating new traits.
+        /// </summary>
         public void SetTraits(GoopTraits traits)
         {
             Traits = traits;
         }
 
+        /// <summary>
+        /// Initializes the goop, and generates random traits if no traits
+        /// exist.
+        /// </summary>
         public void Initialize()
         {
             Stats = new GoopStats(_hungerData, _temperatureData,
