+++ Assets/Scripts/Engine/Goop/Traits/GoopTraitColor.cs
@@ -3,6 +3,9 @@ using UnityEngine;
 
 namespace GoopGame.Engine
 {
+    /// <summary>
+    /// Color implementation of <seealso cref="GoopTrait{T}"/>
+    /// </summary>
     public class GoopTraitColor : GoopTrait<Color>
     {
         public GoopTraitColor(GoopTraitDataColor data)
@@ -24,9 +27,11 @@ namespace GoopGame.Engine
 
         public override Color GenerateCombineValue(Goop goop1, Goop goop2)
         {
+            //Get weights
             GoopWeightStruct struct1 = GoopTraits.GetWeightStruct(goop1);
             GoopWeightStruct struct2 = GoopTraits.GetWeightStruct(goop2);
 
+            //Get color of second goop.
             Color value2 = struct2.GetColor(Type);
 
             return TraitData.GenerateCombineValue(Value, value2, struct1, struct2);
